+++ sketch-compose-core/src/commonTest/kotlin/com/github/panpf/sketch/compose/core/common/test/painter/ComposeBitmapPainterTest.kt
@@ -1,6 +1,7 @@
 package com.github.panpf.sketch.compose.core.common.test.painter
 
 import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.FilterQuality
 import com.github.panpf.sketch.painter.ComposeBitmapPainter
 import com.github.panpf.sketch.painter.asPainter
 import com.github.panpf.sketch.test.utils.createBitmap
@@ -18,6 +19,10 @@ class ComposeBitmapPainterTest {
         assertEquals(
             expected = ComposeBitmapPainter(composeBitmap),
             actual = composeBitmap.asPainter()
+        )
+        assertEquals(
+            expected = ComposeBitmapPainter(composeBitmap, FilterQuality.High),
+            actual = composeBitmap.asPainter(FilterQuality.High)
         )
     }
 
@@ -43,14 +48,19 @@ class ComposeBitmapPainterTest {
         val element1 = ComposeBitmapPainter(composeBitmap1)
         val element11 = ComposeBitmapPainter(composeBitmap1)
         val element2 = ComposeBitmapPainter(composeBitmap2)
+        val element3 = ComposeBitmapPainter(composeBitmap2, FilterQuality.High)
 
         assertEquals(expected = element1, actual = element11)
         assertNotEquals(illegal = element1, actual = element2)
+        assertNotEquals(illegal = element1, actual = element3)
+        assertNotEquals(illegal = element2, actual = element3)
         assertNotEquals(illegal = element1, actual = null as Any?)
         assertNotEquals(illegal = element1, actual = Any())
 
         assertEquals(expected = element1.hashCode(), actual = element11.hashCode())
         assertNotEquals(illegal = element1.hashCode(), actual = element2.hashCode())
+        assertNotEquals(illegal = element1.hashCode(), actual = element3.hashCode())
+        assertNotEquals(illegal = element2.hashCode(), actual = element3.hashCode())
     }
 
     @Test
@@ -58,7 +68,7 @@ class ComposeBitmapPainterTest {
         val composeBitmap = createBitmap(101, 202).toComposeBitmap()
         val element = ComposeBitmapPainter(composeBitmap)
         assertEquals(
-            expected = "ComposeBitmapPainter(bitmap=${composeBitmap.toLogString()})",
+            expected = "ComposeBitmapPainter(bitmap=${composeBitmap.toLogString()}, filterQuality=Low)",
             actual = element.toString()
         )
     }